DON-1112: Refactor, DRY up repeated 0.25 for gift aid factor
